Fixed object not showing up in table view cell.
