+++ Pokedex/Pokedex/Info.plist
